Update dependencies for ember-cli-addon-docs
The latest version of ec-addon-docs relies on the new ember-auto-import
and Webpack 5. Futhermore, it relies on @embroider/util, which I've
downgraded from the manual install so that the dependency linter is
happy.

Additionally, I've removed ember-qunity-assert-helpers, as this seems
like an unmaintained package and the assertions all seem to be in tests
that are either disabled or commented out.